Split off a new utils_selenium.py file from lib/tests/utils.py.
The file needs splitting somehow, and this is a reasonable first step.
